api/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
mysqlclient==2.1.1
pytz==2023.3.post1
sqlparse==0.4.4
